Add code for running prototyping UI
This is a relatively basic Gradio implementation. Complication here has
involved the work to make it auto-reload the code from chat-assistants.
To achieve this that module is installed as editable and then the files
in that directory are added to the watch list. Setting this up involved
establishing you need an absolute path to the watch list files, a
relative path didn't work.

It should be noted that this only expects the python files to be
watched so if we have prompts in YAML they won't be included.