Use a build system to allow REPL usage
Without adding this in I couldn't run `uv run python` and be able to
import the code (e.g. `from chat_prototyping.main import demo`) as it
couldn't find it in src.

This feels a bit of a heavy way to solve this problem, but this is only
a prototype.