Use streaming for Anthropic responses
This changes the sonnet_non_streaming_assistant to be
sonnet_streaming_assistant by utilising streaming from the Anthropic
SDK.

This definitely jumped a difficulty level in testing with lots of
intricacies of python async to be considered.

I found that my previous approach of mocking the underlying HTTP request
for the call to a Bedrock API was really painful to continue with a
streaming approach. This was because you had to break down the response
into each individual object that was streamed back. And if that wasn't
tricky enough it's even harder with Bedrock where SSE is not implemented
and instead there is a proprietary AWS event stream protocol to consider
where every object has a checksum.