Switch to uv workspaces
This allows managing the code in this repo with a single uv lockfile.
I'm not sure if it's the best idea or not, it has the pro that all code
in the repo will use the same version but the disadvantage that we have
to monitor one file changes for all CI runs.